### 2024-12-02 #7546 @editorialbot
[PRE REVIEW]: Pore2Chip: All-in-One Python Tool for Soil Microstructure Analysis and Micromodel Design
**Submitting author:** @aramyxt (<a href="http://orcid.org/0009-0002-7467-7957">Aramy Truong</a>)
**Repository:** https://github.com/EMSL-Computing/Pore2Chip
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
