### 2024-12-02 #7546 @kthyng
@aramyxt I consulted with a license expert from our editorial team about your license and was told:

> It's not an OSI-approved license. It's a mash-up, borrowing phrases from different licenses with unclear implications on compatibility. It could be submitted to OSI for approval, but I expect it would be declined since that would contribute to license proliferation without adding clarity or semantic value.

Note that having an OSI-approved license is a requirement for JOSS to review your submission.
