### 2024-12-04 #7546 @aramyxt
Thank you for the response! I will talk to my team about the license issue.
