### 2024-12-04 9286cfc
Updating license and disclaimer to incorporate inputs from Sara Hunt
