### 2024-12-04 7fb0ee5
The updated license confirms to https://opensource.org/license/bsd-2-clause
