### 2024-12-04 #7546 @maruti-iitm
@kthyng We have removed the disclaimer and have updated the `LICENSE.md`, which now follows the [BSD 2-Clause License](https://opensource.org/license/bsd-2-clause).

Please see the updated license here: [https://github.com/EMSL-Computing/Pore2Chip/blob/main/LICENSE.md](https://github.com/EMSL-Computing/Pore2Chip/blob/main/LICENSE.md)
