### 2024-12-04 33aaf06
Updated paper references to incorporate JOSS editorialbot inputs
