+++ paper/paper.bib
@@ -332,7 +332,8 @@
   number={3},
   pages={vzj2011--0072},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.2136/vzj2011.0072}
 }
 
 @article{anbari2018microfluidic,
@@ -353,7 +354,8 @@
   volume={8},
   pages={6355--6395},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.egyr.2022.04.071}
 }
 
 @article{zhu2022microfluidics,
@@ -364,7 +366,8 @@
   number={2},
   pages={711--731},
   year={2022},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi = {10.1021/acs.est.1c03899}
 }
 
 @article{aralekallu2023development,
@@ -374,7 +377,8 @@
   volume={225},
   pages={111517},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.matdes.2022.111517}
 }
 
 @article{hamamoto2010excluded,
@@ -384,7 +388,8 @@
   volume={46},
   number={6},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2009wr008424}
 }
 
 @article{bailey2017differences,
@@ -394,7 +399,8 @@
   volume={107},
   pages={133--143},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.soilbio.2016.11.025}
 }
 
 @article{davidson2012d,
@@ -416,5 +422,6 @@
   number={16},
   pages={5031--5045},
   year={2018},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  publisher={Copernicus Publications G{\"o}ttingen, Germany},
+  doi = {10.5194/bg-15-5031-2018}
 }
