### 2024-12-04 #7546 @maruti-iitm
> - 10.2136/vzj2011.0072 may be a valid DOI for title: A review of micromodels and their use in two-phase...
> - 10.1016/j.egyr.2022.04.071 may be a valid DOI for title: Fluid–rock interactions and its implications on EO...
> - 10.1021/acs.est.1c03899.s001 may be a valid DOI for title: Microfluidics as an emerging platform for explorin...
> - 10.1016/j.matdes.2022.111517 may be a valid DOI for title: Development of glass-based microfluidic devices: A...
> - 10.1029/2009wr008424 may be a valid DOI for title: Excluded-volume expansion of Archie’s law for gas ...
> - 10.1016/j.soilbio.2016.11.025 may be a valid DOI for title: Differences in soluble organic carbon chemistry in...
> - 10.5194/bg-15-5031-2018 may be a valid DOI for title: Diffusion limitations and Michaelis–Menten kinetic...
