### 2024-12-12 #7546 @kthyng
@maruti-iit, @aramyxt The license still isn't following just one license. It has three clauses and the text above the clauses is mostly borrowed from MIT, not any of the BSD licenses. Is there something preventing you from adopting a single license type?
