### 2024-12-13 #7546 @maruti-iitm
@kthyng, we just got approval to change it to an MIT license.

Here is the updated license text in the repo: https://github.com/EMSL-Computing/Pore2Chip/blob/main/LICENSE.md
