### 2024-12-18 #7546 @maruti-iitm
@kthyng current documentation includes a readme file along with extensive Jupyter Notebooks and video tutorial: https://www.youtube.com/watch?v=_DTLtYBAZSo
