### 2024-12-31 2d5fbb0
Updated Generation Algorithm
Added extra checks to fix problems with pores having a coordination of 0 despite the data set not having 0s. Doing the same with pores with a coordination greater than the max is WIP.
