+++ src/pore2chip/generate.py
@@ -3,7 +3,8 @@ import numpy as np
 import openpnm as op
 import random
 from skimage.morphology import diamond
-
+from itertools import chain
+ 
 
 def generate_network(n1,
                      n2,
@@ -23,7 +24,8 @@ def generate_network(n1,
                      pore_random_shift=0.2,
                      lone_pores=True,
                      center_channel=None,
-                     return_middle_pores=False):
+                     return_middle_pores=False,
+                     sd=0):
     r"""
     Create 2D OpenPNM network with given pore, throat, and coordination 
     information
@@ -52,7 +54,8 @@ def generate_network(n1,
         generated_network : OpenPNM network
     """
 
-    random.seed()
+    random.seed(sd)
+    np.random.seed(sd)
 
     generated_network = op.network.BodyCenteredCubic([n1, n2, 2])
     op.topotools.trim(generated_network, pores=generated_network.pores('zmax'))
@@ -112,29 +115,15 @@ def generate_network(n1,
                           throats=generated_network['throat.all'])
         return generated_network
 
+    # Create list of already visited pores
+    visited = []
+    random.seed(0)
+
     # edit connections to get specified coordination numbers
     for pore_index in range(num_pores):
 
-        if min_pore_diameter is not None:
-            if generated_network['pore.diameter'][
-                    pore_index] > min_pore_diameter:
-                generated_network['pore.diameter'][
-                    pore_index] = min_pore_diameter
-        if max_pore_diameter is not None:
-            if generated_network['pore.diameter'][
-                    pore_index] > max_pore_diameter:
-                generated_network['pore.diameter'][
-                    pore_index] = max_pore_diameter
-
         # Get connected throats
         neighbor_throats = generated_network.find_neighbor_throats(pore_index)
-
-        if min_coordination is not None:
-            if random_coordination[pore_index] < min_coordination:
-                random_coordination[pore_index] = min_coordination
-        if max_coordination is not None:
-            if random_coordination[pore_index] > max_coordination:
-                random_coordination[pore_index] = max_coordination
 
         # If the pore has more throats than what we want...
         if (len(neighbor_throats) > random_coordination[pore_index]
@@ -142,31 +131,73 @@ def generate_network(n1,
             if len(neighbor_throats) == 0:
                 continue
 
-            for throat_index in neighbor_throats:
-                # Disconnect the extra throats
-                if throat_index < len(generated_network['throat.conns']):
-                    op.topotools.trim(generated_network,
-                                      throats=[throat_index])
+            neighbor_pores = generated_network.find_neighbor_pores(
+                pore_index, flatten=True)
+
+            # Number of connections to remove
+            i = len(neighbor_throats) - random_coordination[pore_index]
+            # Number of connections that have been removed
+            j = 0
+            while (i > 0):
+                # If we have made the maximum number of throat removals.
+                # Otherwise, break the loop
+                if j < len(neighbor_pores):
+                    # Pick random neighbor throat
+                    random_throat = np.random.choice(neighbor_throats)
+
+                    # Makes sure throat index is not above number of throats
+                    if random_throat < len(generated_network['throat.conns']):
+
+                        # Connections of the selected throat (tuple of pores the throat connects)
+                        conn = generated_network['throat.conns'][random_throat]
+
+                        # If any of the pores in the tuple have not be visited already
+                        if conn[0] not in visited or conn[1] not in visited:
+                            op.topotools.trim(generated_network,
+                                              throats=[random_throat])
+                            i -= 1  # One less disconnection that needs to be made
+
+                else:
+                    break
+                j += 1  # One more disconnection made
 
         # If the pore has less throats than what we want...
         elif len(neighbor_throats) < random_coordination[pore_index]:
             # Find pores around the vicinity of this pore
[…]
