+++ requirements.txt
@@ -17,4 +17,3 @@ jaxlib
 jax
 optax
 matplotlib
-gstools
