### 2025-01-02 24e6c55
Added Sphinx Doc Files (WIP)
Added files and build system for Sphinx documentation. Work in progress, will add more documentation on functions and quickstart guide.
