+++ docs/build/html/_sources/index.rst.txt
@@ -1,0 +1,23 @@
+.. Pore2Chip documentation master file, created by
+   sphinx-quickstart on Thu Jan  2 09:45:53 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Pore2Chip Documentation
+=======================
+
+**Pore2Chip** is a Python module designed to streamline the process of analyzing X-ray computed tomography (XCT) 
+images of soil and creating 2D micromodel designs based on that analysis. 
+It leverages the power of open-source libraries like OpenPNM, PoreSpy, and drawsvg to extract key information about the soil's 
+porous structure and translate it into a blueprint for microfluidic simulations or physical "lab-on-a-chip" devices developed using additive manufacturing.
+
+.. note::
+
+   This project is under active development.
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   install
+
