+++ .github/workflows/documentation.yml
@@ -16,7 +16,7 @@ jobs:
           pip install sphinx-conestack-theme
       - name: Sphinx build
         run: |
-          sphinx-build doc _build
+          sphinx-build docs _build
       - name: Deploy to GitHub Pages
         uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
