+++ .github/workflows/documentation.yml
@@ -13,12 +13,12 @@ jobs:
       - uses: actions/setup-python@v5
       - name: Install dependencies
         run: |
-          pip install sphinx-conestack-theme
+          pip install sphinx sphinx-conestack-theme myst_parser
       - name: Sphinx build
         run: |
           sphinx-build docs/source _build
       - name: Deploy to GitHub Pages
-        uses: peaceiris/actions-gh-pages@v3
+        uses: emsl-computing/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         with:
           publish_branch: gh-pages
