+++ .github/workflows/documentation.yml
@@ -18,7 +18,7 @@ jobs:
         run: |
           sphinx-build docs/source _build
       - name: Deploy to GitHub Pages
-        uses: emsl-computing/actions-gh-pages@v3
+        uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         with:
           publish_branch: gh-pages
