### 2025-01-09 0b4da2b
Module Reference Docs for filter_im and metrics
Basic autofunction documentation for filter_im and metrics. Need to do the other modules.
