+++ docs/source/coordination_api.rst
@@ -1,0 +1,5 @@
+
+coordination
+============
+
+(WIP)
