### 2025-01-10 20b032b
Edited Paper.md
Trying to cut down wordcount for JOSS paper.
