+++ paper/paper.md
@@ -52,54 +52,55 @@ bibliography: paper.bib
 
 # Summary
 
-The `Pore2Chip` Python package is designed to create two-dimensional micromodels using extracted data from three-dimensional X-ray computed tomography (XCT) images. 
-Micromodels are two-dimensional representations of pore systems that replicate natural porous media, enabling detailed study of fluid dynamics, reactive-transport, and thermal energy at the pore scale [@karadimitriou2012review; @anbari2018microfluidic]. 
-These models are crucial for advancing soil science by providing insights into microbial interactions and chemical processes in Earth and energy systems [@isah2022fluid; @zhu2022microfluidics; @aralekallu2023development].
-This `Pore2Chip` package helps analyze soil structure and function, allowing for the investigation of environmentally significant biogeochemical processes that impact soil organic matter (SOM) decomposition and loss, oxygen concentrations, and nutrient availability in disturbed or managed soils.
-It can produce pore networks that accurately represent soils, convert any 3D pore network into scalable SVG images with precise pore and pore throat sizes, and conduct computational simulations on these images. 
-The extracted data includes characterizations of the pore network using major water retention and flow metrics relevant to porous materials. 
-Key metrics encompass pore size distributions, pore throat size distributions, porosity, tortuosity, and connectivity (pore coordination numbers).
-The software's final output is a 2D lab-on-chip micromodel design representing a soil core or its aggregate. It can be printed using additive manufacturing methods such as 3D printing and photolithography, or subtracted using laser etching.
-These 3D-printed designs accurately depict the XCT-resolved pore networks for soil science applications.
+The `Pore2Chip` Python package is designed to create two-dimensional micromodels using extracted data from three-dimensional X-ray computed tomography 
+(XCT) images. This package helps analyze soil structure and function, allowing for the investigation of environmentally significant biogeochemical processes 
+that impact soil organic matter (SOM) decomposition and loss, oxygen concentrations, and nutrient availability in disturbed or managed soils.  The extracted 
+data includes characterizations of the pore network using major water retention and flow metrics relevant to porous materials. Key metrics encompass pore size 
+distributions, pore throat size distributions, and connectivity (pore coordination numbers). The software's final output is a 2D scalable lab-on-chip SVG 
+design representing a core or aggregate. It can be fabricated with methods such as laser etching, 3D printing, and photolithography.
 
 # Statement of need
 
-The resilience of agricultural and natural landscapes are intrinsically connected to soil structure. 
-Land management (e.g., tillage, grazing, and fire) and associated impacts (e.g., compaction and pore-clogging) along with climate disturbances (e.g., freeze-thaw, flooding, and sea level rise) can transform soil microstructure [@Stoof2016; @Liu2018; @Feng2020; @deOliveira2022; @Rooney2022]. 
-These changes in the soil microstructure determine the flow of water, solutes, and gases, as well as SOM retention, transport, and distribution [@hamamoto2010excluded; @bailey2017differences; @Waring2020]. 
-Simplified, homogeneous pore networks provide innovative demonstrations of how water, solutes, and microbes interact [@Bhattacharjee2022] but need more accurate representations of soil properties. 
-Heterogeneous synthetic habitats are more realistic but time-consuming to design and do not include pore network characteristics, such as pore connectivity or pore throat measurements. 
-Incorporating pore dynamics into soil models, such as SOM degradation, enables dynamic predictions for soil responses under changing pore networks [@davidson2012d; @moyano2018diffusion].
+The resilience of agricultural and natural landscapes is intrinsically connected to soil structure. Land management (e.g., tillage, grazing, and fire) and 
[…]
