### 2025-01-10 28898b0
Fixes to paper.bib
