+++ paper/paper.bib
@@ -404,14 +404,15 @@
 }
 
 @article{davidson2012d,
-  title={The D ual A rrhenius and M ichaelis--M enten kinetics model for decomposition of soil organic matter at hourly to seasonal time scales},
+  title={The Dual Arrhenius and Michaelis--Menten kinetics model for decomposition of soil organic matter at hourly to seasonal time scales},
   author={Davidson, Eric A and Samanta, Sudeep and Caramori, Samantha S and Savage, Kathleen},
   journal={Global change biology},
   volume={18},
   number={1},
   pages={371--384},
   year={2012},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1111/j.1365-2486.2011.02546.x}
 }
 
 @article{moyano2018diffusion,
