### 2025-01-10 cf0c5b5
Further reduce paper.md wordcount
