+++ paper/paper.md
@@ -52,12 +52,11 @@ bibliography: paper.bib
 
 # Summary
 
-The `Pore2Chip` Python package is designed to create two-dimensional micromodels using extracted data from three-dimensional X-ray computed tomography 
+The `Pore2Chip` Python package is designed to create 2D micromodels using extracted data from 3D X-ray computed tomography 
 (XCT) images. This package helps analyze soil structure and function, allowing for the investigation of environmentally significant biogeochemical processes 
-that impact soil organic matter (SOM) decomposition and loss, oxygen concentrations, and nutrient availability in disturbed or managed soils.  The extracted 
-data includes characterizations of the pore network using major water retention and flow metrics relevant to porous materials. Key metrics encompass pore size 
-distributions, pore throat size distributions, and connectivity (pore coordination numbers). The software's final output is a 2D scalable lab-on-chip SVG 
-design representing a core or aggregate. It can be fabricated with methods such as laser etching, 3D printing, and photolithography.
+that impact soil organic matter (SOM) decomposition and loss, oxygen concentrations, and nutrient availability in disturbed or managed soils. Key metrics encompass pore size 
+distributions, pore throat size distributions, and connectivity (pore coordination numbers). The final output is a 2D scalable SVG 
+design representing a core or aggregate. Designs can be fabricated with methods such as laser etching, 3D printing, and photolithography.
 
 # Statement of need
 
@@ -67,15 +66,15 @@ microstructure [@Stoof2016; @Liu2018; @Feng2020; @deOliveira2022; @Rooney2022]. 
 determine the flow of water, solutes, and gasses as well as SOM retention, transport, and distribution 
 [@hamamoto2010excluded; @bailey2017differences; @Waring2020]. Simplified, homogeneous pore networks provide innovative 
 demonstrations of how water, solutes, and microbes interact [@Bhattacharjee2022] but need more accurate representations of soil properties. 
-Heterogeneous synthetic habitats are more realistic but time-consuming to design and do not include pore network characteristics, such as pore 
-connectivity or pore throat measurements. Incorporating pore dynamics into soil models such as SOM degradation enables dynamic predictions for soil 
+Creating realistic heterogeneous habitats is time-consuming and does not include pore network characteristics, such as pore 
+connectivity. Incorporating pore dynamics into soil models such as SOM degradation enables dynamic predictions for soil 
 responses under changing pore networks [@davidson2012d; @moyano2018diffusion].
 
-Currently, there is no software available to seamlessly provide various micromodel designs that researchers can test and validate with minimal computational 
-cost (Dentz et al., 2023; Oostrom et al., 2014). `Pore2Chip` allows us to overcome this barrier by providing the intended users, such as earth scientists and 
-lab-on-chip instrument specialists, with easy-to-use research software for lab-on-chip designs. Specifically, the Pore2Chip-based data worth analysis of 
-high-resolution XCT images allows us to fill this experimental design gap by allowing the users to build a representative quasi-2D pore network along with 
-first-order, fast, and reasonably accurate flow models that can be linked with experiments. These `Pore2Chip` flow models are built using recent advances in 
+The need for software that can generate various micromodel designs that researchers can test and validate with minimal computational 
+cost [@Dentz2023; @Oostrom2014] is increasing. `Pore2Chip` allows this functionality by providing the intended users, such as earth scientists and 
+lab-on-chip instrument specialists, with easy-to-use research software for lab-on-chip designs. Specifically, the Pore2Chip-based data worth analysis allows researchers to 
+fill this experimental design gap by enabling the ability to build a representative quasi-2D pore network along with 
+first-order, fast, and reasonably accurate flow models that can be linked with experiments. These flow models are built using recent advances in 
[…]
