### 2025-01-10 #7546 @aramyxt
@cheginit @kthyng @maruti-iitm

Hello, I have reduced the wordcount of the paper to 935 words and also fixed the strange formatting error with one of the references. Please reach out to me if there are any more changes that I need to make to be compliant with JOSS formatting requirements. Thank you!
