### 2025-01-13 #7546 @cheginit
👋🏼 @maruti-iitm @aramyxt thanks for addressing the issue with the paper. I went through the list of potential reviewers that you provided, but unfortunately none of them appear to work in a related field and have experience with Python. Also, please note that they should have a GitHub account and have no the conflict of interest, e.g., working in the same institution as yours. You can review JOSS's policy on this [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). You can use JOSS' [tool](https://reviewers.joss.theoj.org/reviewers) to search for reviewers in addition to other people in your own network that have the required expertise.
