### 2025-01-13 #7546 @maruti-iitm
Hi @cheginit, Thank you so much! Happy New Year! The potential reviewers suggested here are not co-authors (in any of our previous research work/papers): https://github.com/openjournals/joss-reviews/issues/7546#issuecomment-2518749081

They have experience with software-based experimental design and Python programming (e.g., Pratanu Roy and Habiba Akter). @aramyxt can correct me if I am wrong.

Sure, moving forward, we will use the JOSS tool to search for additional reviewers and get back to you as soon as we can. Is that okay?
