### 2025-01-14 #7546 @cheginit
@maruti-iitm Thanks, happy new year to you too!

> The potential reviewers suggested here are not co-authors

I was referring to Jaysaval, who, based on his GitHub profile, works at PNNL where you are affiliated with.

> we will use the JOSS tool to search for additional reviewers and get back to you as soon as we can. Is that okay?

Using the search tool is just an option that can be helpful, you don't have to use it. So, please feel free to explore different sources for finding suitable reviewers. As long as you provide me their GitHub handle (without using `@` to avoid premature pinging), it's fine.
