### 2025-01-14 084fb81
Temporary fix to Porespy/Scikit-Image deprecation
The latest version of Porespy has not merged their fix for the deprecated scikit-image "skeletonize_3D" function into the main development branch yet, so this commit is a workaround for the sake of importing and Sphinx documentation.
