### 2025-01-15 81bd1a2
Restored requirements.txt
Porespy has merged its scikit-image fixes, so the workaround requirements are no longer necessary.
