### 2025-01-15 8fd181e
Update documentation.yml
Testing to fix Sphinx build hanging
