+++ .github/workflows/documentation.yml
@@ -13,7 +13,7 @@ jobs:
       - uses: actions/setup-python@v5
       - name: Install dependencies
         run: |
-          pip install sphinx sphinx-conestack-theme myst_parser
+          pip install sphinx sphinx-conestack-theme
           pip install -e .
       - name: Sphinx build
         run: |
