### 2025-01-20 #7546 @aramyxt
@cheginit @maruti-iitm
Hello all,

After looking through potential reviewers using the JOSS tool, I came up with basically the same list as Maruti. I will add a couple suggestions here:
- filippi (Jean-Baptiste Filippi)
- gartavanis (Georgios Artavanis)
