### 2025-01-27 ce4f973
Added Documentation for coordination and generate
Also added links to github and pypi page.
