+++ docs/source/generate_api.rst
@@ -1,5 +1,13 @@
 
-generate
-========
+**generate**
+============
 
-(WIP)
+The ``generate`` module contains a function to generate an ``openpnm.models.network`` object that 
+contains a 2D network based on input pore data (see ``metrics`` for more info on extracting data).
+
+----
+
+generate_network()
+------------------
+
+.. autofunction:: pore2chip.generate.generate_network
