### 2025-02-05 #7546 @lukashof
Thank you for the opportunity, @cheginit! Unfortunately, the application seems to be outside my area of expertise and I would like to decline. I hope you find someone quickly!
