### 2025-02-26 #7546 @aramyxt
Hello @cheginit,
How is the status of the paper? Is there anything else that needs to be done on our end?
