+++ paper/paper.bib
@@ -344,7 +344,8 @@
   number={18},
   pages={1703575},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1002/smll.201703575}
 }
 
 @article{isah2022fluid,
