### 2025-04-23 #8052 @shahram444
Installation

Does installation proceed as outlined in the documentation?
Yes, the installation using pip install pore2chip proceeds correctly on a clean Python 3.9 Conda environment. The required dependencies such as porespy, opencv-python, openpnm, and others are installed automatically via PyPI.

However, two additional setup notes are important for successful usage, especially in the Jupyter or Spyder environments:
1. reportlab version compatibility

    Issue: The latest version of reportlab (≥4.0) is not compatible with svglib, which is used for converting SVG to raster image formats (e.g., PNG).

    Fix: Downgrade reportlab to version 3.6.13:

                 pip install reportlab==3.6.13

2. %matplotlib widget causes backend errors

    Issue: Several notebooks start with %matplotlib widget, which fails in non-notebook environments (e.g., Spyder, some VSCode setups).

    Fix:  %matplotlib inline

3. TqdmWarning: IProgress not found

    Issue: This warning appears when using progress bars in notebooks:

TqdmWarning: IProgress not found. Please update jupyter and ipywidgets.

Functionality

Have the functional claims of the software been confirmed?
Yes. All primary functionalities described in the paper and README were successfully tested using the provided examples. These functionalities include: XCT image filtering, pore-throat metric extraction, synthetic network generation, finite difference flow simulation, PINN-based prediction, VTK export, and data visualization. All these tasks were tested in their respective examples and validated.

here is a summary of each example, including changes I made and inline comments

Example 1 – Micromodel Generation from 50×50 XCT Image

What worked:

    XCT image preprocessing using filter_im.read_and_filter()

    Metric extraction via metrics.get_pore_sizes() and metrics.get_throat_sizes()

    Network generation via generate.generate_network()

    SVG and PNG export with export.design2svg()

# %matplotlib widget

#I changes with:
%matplotlib inline  # Ensures stable rendering in Jupyter and IDEs like Spyder

# SOLUTION:
pip install reportlab==3.6.13  # Downgrades reportlab to match svglib compatibility

Example 2 – Micromodel Generation from 100×100 XCT Image

What worked:

    Larger image processed and analyzed without issue.

# I changed to:
fig, ax = plt.subplots(figsize=(10, 5))  # Create figure and axis
op.visualization.plot_connections(network, ax=ax)
op.visualization.plot_coordinates(network, ax=ax)
plt.title("Generated Micromodel Network")

Example 3 – Metric Visualization

What worked:

    Feret diameter and coordination number plots

    XCT-derived porosity metrics

Fixes I made:

# SOLUTION:
%matplotlib inline

Example 4 – Tortuosity & Drainage (PoreSpy)

What worked:

    Tortuosity simulation and capillary pressure drainage curves

# ADD THIS BEFORE CALLING io.img2vtk
os.makedirs('example_outputs', exist_ok=True)  # Ensures output folder exists

Example 5 – 2D Flow Simulation (Finite Difference)

What worked:

    Flow solved across heterogeneous permeability field

# WITH THIS:
from scipy.sparse import lil_matrix
A = lil_matrix((nx * ny, nx * ny))  # Memory-efficient sparse matrix

# WITH THIS:
from scipy.sparse.linalg import spsolve
A_csr = A.tocsr()  # Convert to compressed sparse format
P_flat = spsolve(A_csr, b)

Example 6 – Masked Pressure Visualization

What worked:

    Masked pressure overlaid on pore-solid map

I made change inside plot_2d_pressure_distribution_masked() in src/flow/plotting_results.py:

# I added this to end of function:
plt.show()  # Forces figure to appear in non-interactive backends

Example 7 – PINN Flow Simulation

What worked:

    Neural net trained using Optax, PINN predicted pressure field correctly

Fixes made in plot_boundary_conditions() in src/flow/plotting_results.py:

# This does not link the colorbar to any specific scatter plot, causing issues:
cbar = plt.colorbar(mpl.cm.ScalarMappable(norm=mpl.colors.Normalize(vmin=0, vmax=2), cmap=cm.jet), aspect=30)

Corrected Version

# Plot the first boundary condition and assign its handle to 'sc'
sc = plt.scatter(x_b1, y_b1, c=bc_1, marker='x', vmin=0, vmax=2,
                 label='P[0,y] = 0', cmap=cm.jet, s=10)

# Plot remaining boundary conditions (no need to assign handles for colorbar)
plt.scatter(x_b2, y_b2, c=bc_2, marker='^', vmin=0, vmax=2,
            label='P[1,y] = y', cmap=cm.jet, s=10)

plt.scatter(x_b3, y_b3, c=bc_3, marker='*', vmin=0, vmax=2,
            label='$\\partial P/\\partial y[x,0] = 0$', cmap=cm.jet, s=10)

plt.scatter(x_b4, y_b4, c=bc_4, marker='o', vmin=0, vmax=2,
            label='$\\partial P/\\partial y[x,1] = 0$', cmap=cm.jet, s=10)

# Plot collocation points in black (no colorbar needed)
plt.scatter(x_c, y_c, c='k', marker='.', alpha=0.5,
            label='Collocation points', s=5)

# Now correctly link the colorbar to the first scatter plot (sc)
cbar = plt.colorbar(sc, aspect=30)  # Use 'sc' as mappable
cbar.set_label('Pressure (kPa)')

Performance

The Pore2Chip package doesn’t claim any specific performance benchmarks in its documentation or paper, but I still tested it across several examples to get a sense of how it behaves in practice — particularly with respect to memory usage, runtime, and general usability on a standard machine.
General Observations:

    I ran all examples on a regular laptop (16 GB RAM, no GPU). Everything worked fine, including the larger simulations.

    The examples involving XCT datasets sized at 50×50, 100×100, and even 1200×1200 (in the finite difference and PINN cases) executed without major issues.

    PINN training and VTK export steps took the most time, but nothing was unreasonable — everything remained within manageable limits for local runs.
