### 2025-04-23 #8052 @shahram444
Documentation Review
1. Statement of Need
The README does a solid job explaining what the software is for: it's designed for working on micromodels, XCT image analysis, pore network extraction, and flow simulations — mostly in porous media research. It clearly targets a scientific audience, and it's easy to see where it fits into the current ecosystem (bridging between XCT data and physics-based models, including PINNs). Nothing fancy, but very clear.
2. Installation Instructions
Basic install using pip install pore2chip works — no complaints there. The requirements.txt is included and functional. But a few manual tweaks were needed:
Issue: reportlab ≥ 4.0 breaks SVG-to-PNG conversion because it’s incompatible with svglib.
Fix: We pinned reportlab==3.6.13
Issue: pyfastnoisesimd is required for porespy.generators.fractal_noise but isn't auto-installed.
Fix: Added manually with pip install pyfastnoisesimd
Issue: %matplotlib widget errors in Spyder / some notebooks
Fix: Replaced with %matplotlib inline or just used plt.show()
3. Example Usage
All the examples work (after small tweaks). They cover realistic workflows: image filtering, network generation, metric extraction, flow simulations, and training PINNs. I ran everything from examples/ and it’s really helpful that it shows the whole pipeline step by step.
4. Functionality Documentation
Most functions have docstrings that explain inputs/outputs, especially in generate.py, metrics.py, and flow/pinn_utilities.py. But a few modules (like filter_im.py and export.py) could use more detailed comments — sometimes it's not clear what optional arguments are expected.
5. Automated Tests
There’s a full tests/ folder and I ran every script inside. They actually cover a lot:
test_coordination.py: This Tests both 2D and 3D coordination number calculations using OpenPNM and PoreSpy images. Everything passed without issues. No changes needed.
test_export.py: Verifies the SVG and DXF export functionality. Initially raised an IndexError due to the use of Path(__file__).resolve().parents[2]. I fixed it by switching to Path(os.getcwd()).resolve().parent, which worked correctly in all environments.
test_filter_im.py: Tests single-image and image-stack filtering (cropping, thresholding, Gaussian blur). It passed after removing plt.waitforbuttonpress() — that call was causing backend issues in Spyder and other non-interactive environments.
test_generate.py: Generates networks using both random property sampling and probability density functions. I tested different grid sizes, center channel options, and coordination number settings. All configurations ran correctly — no modifications required.
test_metrics.py: Tests all three diameter extraction methods (extract_diameters, extract_diameters_alt, and extract_diameters2) as well as feret_diameter. I used both synthetic 2D ellipses and 3D cylinders as input. Everything passed and produced meaningful results.
test_train_pinn.py: This was the most comprehensive. It tests train_PINN under multiple conditions: output shape, edge cases, exception handling, different optimizers, loss functions, and architectures. All 9 test cases passed. The function handled bad inputs correctly and produced stable results for all tested setups.
Changes we made to get all tests working:
# old:
mod_path = Path("__file__").resolve().parents[2]
# new:
mod_path = Path(os.getcwd()).resolve().parent  # safer in any IDE or terminal

# old:
plt.waitforbuttonpress()
# new:
plt.show()  # just shows the plot and continues, no interactive issues
# Dependency fix:
pip install pyfastnoisesimd
pip install reportlab==3.6.13
6. Community Guidelines
The repository includes basic author and license information (MIT) in the `README.md`, and it's clear that the project is open for use and collaboration. However, there are no structured guidelines for contributing, reporting issues, or requesting support. I suggest adding a short "Contributing" section or a separate `CONTRIBUTING.md` that outlines how to open issues, submit pull requests, and ask for help. This would make it easier for others to get involved and ensure consistency in community contributions.
