### 2025-04-28 #8052 @mlmamud
> # I am working to address the following:
> 3. 2. %matplotlib widget causes backend errors
>    Issue: Several notebooks start with %matplotlib widget, which fails in non-notebook environments (e.g., Spyder, some VSCode setups).
>    Fix:  %matplotlib inline
> 3. TqdmWarning: IProgress not found
>    Issue: This warning appears when using progress bars in notebooks:
>
> TqdmWarning: IProgress not found. Please update jupyter and ipywidgets.
>
> Functionality
>
> Example 5 – 2D Flow Simulation (Finite Difference)
>
> What worked:
>
> ```
> Flow solved across a heterogeneous permeability field
> ```
>
> # FIX: Dense matrix caused memory error (11.9 GB)
> # I replaced this:
> # A = np.zeros((nx * ny, nx * ny))
> # WITH THIS:
> from scipy.sparse import lil_matrix A = lil_matrix((nx * ny, nx * ny)) # Memory-efficient sparse matrix
>
> # FIX: Dense solver incompatible with sparse matrix
> # I replaced this:
> # P_flat = np.linalg.solve(A, b)
> # WITH THIS:
> from scipy.sparse.linalg import spsolve A_csr = A.tocsr() # Convert to compressed sparse format P_flat = spsolve(A_csr, b)
>
> Example 6 – Masked Pressure Visualization
>
> What worked:
>
> ```
> Masked pressure overlaid on pore-solid map
> ```
>
> I made change inside plot_2d_pressure_distribution_masked() in src/flow/plotting_results.py:
>
> # FIX: Plot did not render in Spyder or some notebooks
> # I added this to end of function:
> plt.show() # Forces figure to appear in non-interactive backends
>
> Example 7 – PINN Flow Simulation
>
> What worked:
>
> ```
> Neural net trained using Optax, PINN predicted pressure field correctly
> ```
>
> Fixes made in plot_boundary_conditions() in src/flow/plotting_results.py:
>
> # This does not link the colorbar to any specific scatter plot, causing issues:
> cbar = plt.colorbar(mpl.cm.ScalarMappable(norm=mpl.colors.Normalize(vmin=0, vmax=2), cmap=cm.jet), aspect=30)
>
> Corrected Version
>
> # FIX: Link the colorbar to an actual scatter plot object
> # Plot the first boundary condition and assign its handle to 'sc'
> sc = plt.scatter(x_b1, y_b1, c=bc_1, marker='x', vmin=0, vmax=2, label='P[0,y] = 0', cmap=cm.jet, s=10)
>
> # Plot remaining boundary conditions (no need to assign handles for colorbar)
> plt.scatter(x_b2, y_b2, c=bc_2, marker='^', vmin=0, vmax=2, label='P[1,y] = y', cmap=cm.jet, s=10)
>
> plt.scatter(x_b3, y_b3, c=bc_3, marker='*', vmin=0, vmax=2, label='$\partial P/\partial y[x,0] = 0$', cmap=cm.jet, s=10)
>
> plt.scatter(x_b4, y_b4, c=bc_4, marker='o', vmin=0, vmax=2, label='$\partial P/\partial y[x,1] = 0$', cmap=cm.jet, s=10)
>
> # Plot collocation points in black (no colorbar needed)
> plt.scatter(x_c, y_c, c='k', marker='.', alpha=0.5, label='Collocation points', s=5)
>
> # Now correctly link the colorbar to the first scatter plot (sc)
> cbar = plt.colorbar(sc, aspect=30) # Use 'sc' as mappable cbar.set_label('Pressure (kPa)')
