+++ src/flow/plotting_results.py
@@ -79,76 +79,9 @@ def plot_2d_permeability_seg(X, Y, perm_field, results_dir):
     plt.savefig(results_dir + filename, dpi=dpi)
 
 
-def plot_boundary_conditions(x_b1, y_b1, bc_1, x_b2, y_b2, bc_2, x_b3, y_b3,
-                             bc_3, x_b4, y_b4, bc_4, x_c, y_c, results_dir):
-    plt.figure(figsize=(width, height), dpi=dpi)
-    s = 10
-    plt.scatter(x_b1,
-                y_b1,
-                c=bc_1,
-                marker='x',
-                vmin=0,
-                vmax=2,
-                label='P[0,y] = 0',
-                cmap=cm.jet,
-                s=s)
-    plt.scatter(x_b2,
-                y_b2,
-                c=bc_2,
-                marker='^',
-                vmin=0,
-                vmax=2,
-                label='P[1,y] = y',
-                cmap=cm.jet,
-                s=s)
-    plt.scatter(x_b3,
-                y_b3,
-                c=bc_3,
-                marker='*',
-                vmin=0,
-                vmax=2,
-                label='$\\partial P/\\partial y[x,0] = 0$',
-                cmap=cm.jet,
-                s=s)
-    plt.scatter(x_b4,
-                y_b4,
-                c=bc_4,
-                marker='o',
-                vmin=0,
-                vmax=2,
-                label='$\\partial P/\\partial y[x,1] = 0$',
-                cmap=cm.jet,
-                s=s)
-    s = 5
-    plt.scatter(x_c,
-                y_c,
-                c='k',
-                marker='.',
-                alpha=0.5,
-                label='Collocation points',
-                s=s)
-    # Labels and colorbar
-    plt.xlabel('$X$')
-    plt.ylabel('$Y$')
-    cbar = plt.colorbar(mpl.cm.ScalarMappable(norm=mpl.colors.Normalize(
-        vmin=0, vmax=2),
-                                              cmap=cm.jet),
-                        aspect=30)
-    cbar.set_label('Pressure (kPa)')
-    plt.tight_layout()
-    figure_name = 'bcs_collocs'
-    plt.savefig(results_dir + figure_name + figure_format,
-                dpi=dpi,
-                format=figure_format.strip('.'),
-                bbox_inches='tight')
-    plt.show()
-
-
-
-
-def plot_boundary_conditions2(x_b1, y_b1, bc_1, x_b2, y_b2, bc_2,
-                             x_b3, y_b3, bc_3, x_b4, y_b4, bc_4,
-                             x_c, y_c, results_dir):
+def plot_boundary_conditions(x_b1, y_b1, bc_1, x_b2, y_b2, bc_2,
+                              x_b3, y_b3, bc_3, x_b4, y_b4, bc_4,
+                              x_c, y_c, results_dir):
     plt.figure(figsize=(width, height), dpi=dpi)
 
     # Compute vmin and vmax across all boundary condition values
@@ -156,31 +89,34 @@ def plot_boundary_conditions2(x_b1, y_b1, bc_1, x_b2, y_b2, bc_2,
     vmin = np.min(all_bc_values)
     vmax = np.max(all_bc_values)
 
-    s = 10
+    s = 3
 
     # Plot the first boundary condition and link it to the colorbar
-    sc = plt.scatter(x_b1, y_b1, c=bc_1, marker='x', vmin=0, vmax=2,
-                     label='P[0,y] = 0', cmap=cm.jet, s=s)
+    sc = plt.scatter(x_b1, y_b1, c=bc_1, marker='x', vmin=vmin, vmax=vmax,
+                     label='P[0,y]', cmap=cm.jet, s=s)
 
     # Remaining boundary conditions
[…]
