### 2025-06-18 a6ec949
Update to readme bullets related to modex loop
