### 2025-06-18 39ad191
Update to readme bullets related to modex loop; reverting back
