### 2025-06-18 cb3511e
Update to readme bullets related to modex loop; to make it consistent
