### 2025-06-18 efdf6e3
Update to readme bullets related to modex loop; to make it consistent across modex bullets
