### 2025-06-18 e1f12df
Update filter_im.py
Added sorted() to functions using os.listdir to ensure image files are loaded in alphabetical order.
