+++ _sources/metrics_api.rst.txt
@@ -1,0 +1,59 @@
+
+**metrics**
+===========
+
+The ``metrics`` module contains functions that extract pore data from segmented images.
+
+----
+
+feret_diameter()
+----------------
+
+.. autofunction:: pore2chip.metrics.feret_diameter
+
+----
+
+feret_diameter_list()
+---------------------
+
+.. autofunction:: pore2chip.metrics.feret_diameter_list
+
+----
+
+extract_diameters()
+-------------------
+
+.. autofunction:: pore2chip.metrics.extract_diameters
+
+----
+
+extract_diameters2()
+--------------------
+
+.. autofunction:: pore2chip.metrics.extract_diameters2
+
+----
+
+extract_diameters_alt()
+-----------------------
+
+.. autofunction:: pore2chip.metrics.extract_diameters_alt
+
+----
+
+get_probability_density()
+-------------------------
+
+.. autofunction:: pore2chip.metrics.get_probability_density
+
+----
+
+get_percent_probability()
+-------------------------
+
+.. autofunction:: pore2chip.metrics.get_percent_probability
+
+
+.. note::
+
+   This project is under active development.
