### 2025-06-26 #8052 @jgostick
The readme makes a lot more sense now.  It's fine to include the graphic showing how pore2chip fits in with the rest of your ecosystem.
