### 2025-06-26 #8052 @jgostick
Suggested corrections to paper:
- [ ] The work 'worth' seems out of place here: "Specifically, the Pore2Chip-based data worth analysis"
- [ ] This sentence seems odd: "The generation function can also work with data extracted by other means if it is an array of values that Python can read."  Should this be: "The generation function can also work with data extracted by other means provided it is compatible with Python"?
- [ ] I am not sure what the comparison to epyc is for?  epyc is about running many experiments/simulations in parallel.  You say pore2chip is better, but not how or why, and it's also not clear how a package that is about making microfluidic chips can be better than a package that is made for purpose.
- [ ] I generally find the wording of the paper to be unnecessarily verbose. Clear is better than fancy.
