### 2025-07-04 1b01f14
addressed the issues 17, 18, and 19 of JOSS review
