### 2025-07-15 #8052 @aramyxt
Hi @jgostick , thank you for your comments!
I will try to address them in this post, but if I miss anything, please tag me. I apologize for the delays, as most of our team members have been very busy as of late.

**Repository Changes:**

- For Mac installation issues, I will refer to your message here: https://github.com/openjournals/joss-reviews/issues/8052#issuecomment-3011115205 Thank you for helping with this! We intend Pore2Chip to be available on all platforms, but Mac systems were our blind spot. If this is okay, I would like to merge this change into our main branch.
- For the example usage, we have a link to our Github docs page in the “Getting Started” section of our readme.md. There might be a link that takes you to the EMSL page for Pore2Chip, which I would like to keep since it shows how the software can and has been used in real research applications. That being said, we will probably put the documentation and examples front and center in our readme.
- As for the test suites, I will work on actually using ```pytest``` to test outputs. Before, I was not used to using testing packages, but now that I’m more comfortable with it, I will work to convert existing test scripts, following the example scipt you posted here (thank you!): https://github.com/openjournals/joss-reviews/issues/8052#issuecomment-3010095986
- I will also look into providing install options for people who want to use ```uv```. This is still new to me, but I will look into it!
- We have a branch called ```joss-edits``` at the moment, which includes some changes to resolve our Github repo issues. I will merge them into the main branch and work on improving the test suites in a separate branch.

**Paper Changes:**
Here are some replies for the paper itself in a Q&A format:
Q: The work 'worth' seems out of place here: "Specifically, the Pore2Chip-based data worth analysis"
A: We agree, it is better to just have “Specifically, the Pore2Chip-based value of information analysis was used to evaluate how XCT images  influence micromodel design decisions”

Q: This sentence seems odd: "The generation function can also work with data extracted by other means if it is an array of values that Python can read." Should this be: "The generation function can also work with data extracted by other means provided it is compatible with Python"?
A: By this, we mean that the data of interest (pore size distributions, coordination numbers, etc.) does not need to be extracted using the Porespy module that is built into Pore2Chip functions. The only input for the generate function is a list/array of the required values. These values could be imported from a CSV file from software like Aviso or Dragonfly. I think we can definitely re-word this so that it reads more clearly.

Q: I am not sure what the comparison to epyc is for? epyc is about running many experiments/simulations in parallel. You say pore2chip is better, but not how or why, and it's also not clear how a package that is about making microfluidic chips can be better than a package that is made for purpose.
A: The comparison was not meant to imply direct equivalence. Rather, the intent was to highlight that `Pore2Chip` provides domain-specific automation and data management of XCT images. Specifically, `Pore2Chip` is tailored for a particular class of experimental design, namely, microfluidic chips derived from porous media like soil or rock.
While `epyc` excels in managing and running a large number of generic experiments/simulations in parallel and not for lab-on-chip experiments, Pore2Chip is built for a more targeted pipeline: extracting meaningful geometric data from 3D XCT images, analyzing pore-scale structures, and translating that into chip-ready SVG designs. It includes tooling for geometric processing, morphological filtering, micromodel design generation, and quick fluid-flow analysis, all under a reproducible, extensible Python-based workflow. This makes it purpose-built for a high-impact use case, especially in experimental geosciences and microfluidics.
So to clarify: the intention wasn't to say Pore2Chip is better overall, but rather that it is better suited for a specific domain and workflow that is not the target of epyc.

Q: I generally find the wording of the paper to be unnecessarily verbose. Clear is better than fancy.
A: We gave our best effort to convey Pore2Chip and its multiple facets in the experimental world within the JOSS word limit. Please let us know which sections require better clarity. We are expanding it to critical mineral applications, allowing users familiar with biomining to utilize this capability.

I hope I didn’t miss anything, and please ping me if something needs more clarification. These responses were drafted by me and my team, so please also ping @mlmamud and @maruti-iitm  if there are more changes that need to be addressed. Thank you again and thank you @cheginit for helping!
