+++ _sources/module_ref.rst.txt
@@ -1,0 +1,15 @@
+
+Module Reference
+================
+
+Here are the references for all Pore2Chip's modules:
+
+.. toctree::
+   :maxdepth: 2
+
+   filter_im_api
+   metrics_api
+   coordination_api
+   generate_api
+   export_api
+   io_api
