### 2025-07-15 #8052 @maruti-iitm
Hi @jgostick, thank you for your input! We have incorporated your feedback and pushed it to `main` branch as @aramyxt mentioned.

Also, we (@aramyxt @mlmamud @maruti-iitm ) have responded to the GitHub issues (in `Pore2Chip` repository) and tagged you. Please let me know if you require any additional information.
