### 2025-08-11 #8052 @aramyxt
Thank you @cheginit !

I will work on releasing the new version on our Github, as well as updating the PyPi repo for Pore2Chip. I will work with @maruti-iitm on getting the repo archived on Zenodo by the end of this week. Please reach out if you have any more questions. Thank you!
