### 2025-08-14 df4ae04
Misc JOSS Edits
Preparing for 0.2.0 release. More edits to come.
