+++ docs/source/export_api.rst
@@ -1,5 +1,24 @@
 
-export
-======
+**export**
+==========
 
-(WIP)
+The ``export`` module contains functions to draw a visual micromodel mask from an OpenPNM network generated
+from the ``generate`` module (see ``generate`` for more info on generating a network).
+
+
+network2svg()
+-------------
+
+.. autofunction:: pore2chip.export.network2svg
+
+----
+
+network2dxf()
+-------------
+
+.. autofunction:: pore2chip.export.network2dxf
+
+
+.. note::
+
+   This project is under active development.
