+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pore2chip"
-version = "0.1.1"
+version = "0.2.0"
 authors = [
   { name="Aramy Truong", email="aramy.truong@pnnl.gov" },
   { name="Maruti Mudunuru", email="maruti@pnnl.gov" },
