### 2025-08-15 #8052 @maruti-iitm
Hi @cheginit, we (@maruti-iitm and @aramyxt) are almost done with the Zenodo and are releasing a new version. We are double-checking the repo, going through the internal information release process, and will be done by the end of Monday/Tuesday morning (Aug-18-2025/Aug-19-2025). Is that okay?
