### 2025-08-17 a555b73
Updated to JOSS paper acceptance and figures
