### 2025-08-17 2b093b8
Updated to JOSS paper
