### 2025-08-17 8615d49
Update to readme for JOSS paper acceptance
