### 2025-08-18 e61648e
More JOSS Edits
More edits for JOSS publication, as well as starting implementation of pytest to the test suites.
