+++ tests/test_coordination.py
@@ -4,14 +4,14 @@ import time
 import openpnm as op
 import porespy as ps
 from pathlib import Path
+import pytest
+import numpy as np
 
 # Uncomment and modify these lines if the pore2chip package is not in the PYTHONPATH
 #mod_path = Path("__file__").resolve().parents[2]
 #sys.path.append(os.path.abspath(mod_path))
 
 from pore2chip.coordination import coordination_nums_2D, coordination_nums_3D
-#from pore2chip.src.pore2chip.export import coordination_nums_3D, coordination_nums_2D
-
 
 def generate_network():
     """
@@ -22,64 +22,75 @@ def generate_network():
     """
     network = op.network.Cubic([3, 3])  # Create a 3x3 cubic network
     network.add_model_collection(op.models.collections.geometry.
-                                 spheres_and_cylinders)  # Add geometry models
+                                spheres_and_cylinders)  # Add geometry models
     network.regenerate_models()  # Generate network properties
     print(network)  # For debugging purposes
     return network
 
+class TestCoordination():
 
-def test_coordination_nums_3D(network):
-    """
-    Test coordination_nums_2D function given an OpenPNM network.
+    def test_coordination_nums_3D(self):
+        """
+        Test coordination_nums_2D function given an OpenPNM network.
     
-    Args:
-        network (dict): The network structure and coordination numbers (e.g., op.network.Cubic).
+        Args:
+            network (dict): The network structure and coordination numbers (e.g., op.network.Cubic).
 
-    Returns:
-        coordination_numbers (ndarray): An array of coordination numbers for each pore in the network..
-    """
-    coordination_numbers = coordination_nums_3D(pn=network)
-    print(coordination_numbers)  # For debugging purposes
-    print(coordination_numbers.mean())  # For debugging purposes
-    print(coordination_numbers.shape)  # For debugging purposes
-    return coordination_numbers
+        Returns:
+            coordination_numbers (ndarray): An array of coordination numbers for each pore in the network..
+        """
+        network = generate_network()
+        coordination_numbers = coordination_nums_3D(pn=network)
+        assert coordination_numbers.any() == np.array([2,3,2,3,4,3,2,3,2],dtype=int).any()
+        #print(coordination_numbers)  # For debugging purposes
+        #print(coordination_numbers.mean())  # For debugging purposes
+        #print(coordination_numbers.shape)  # For debugging purposes
+        #return coordination_numbers
 
 
-def test_coordination_nums_2D(img_list):
-    """
-    Test coordination_nums_2D function given a list of images.
+    def test_coordination_nums_2D(self):
+        """
+        Test coordination_nums_2D function given a list of images.
     
-    Args:
-        img_list (ndarray): 3D image list
+        Args:
+            img_list (ndarray): 3D image list
 
-    Returns:
-        coordination_numbers (ndarray): An array of coordination numbers for each pore in the network..
-    """
-    coordination_numbers = coordination_nums_2D(img_list)
-    print(coordination_numbers)
-    print(coordination_numbers.mean())
-    print(coordination_numbers.shape)
-    return coordination_numbers
-
-
-def main():
-    """
-    Main function to generate a network, test coordination number computations on 3D networks
-    and 2D image slices.
-    """
-
-    # Generate network
-    network = generate_network()
-
-    # Test coordination 3D on generated network
[…]
