### 2025-08-21 #8052 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
