### 2025-08-26 #8052 @kthyng
@aramyxt Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "archie's" but please thoroughly check.
