### 2025-08-27 750ca9d
Update to paper.bib incorporating editor's feedback
