+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @article{Adem1984,
   author = {H. H. Adem and J. M. Tisdall and P. Willoughby},
-  title = {Tillage management changes the size-distribution of aggregates and macro-structure of soils used for irrigated row-crops},
+  title = {{Tillage} management changes the size-distribution of aggregates and macro-structure of soils used for irrigated row-crops},
   journal = {Soil and Tillage Research},
   volume = {4},
   number = {6},
@@ -10,8 +10,8 @@
 }
 
 @article{Bhattacharjee2022,
-  author = {A. Bhattacharjee and O. Qafoku and J. A. Richardson and L. N. Anderson and K. Schwarz et al.},
-  title = {A Mineral-Doped Micromodel Platform Demonstrates Fungal Bridging of Carbon Hot Spots and Hyphal Transport of Mineral-Derived Nutrients},
+  author = {A. Bhattacharjee and O. Qafoku and J. A. Richardson and L. N. Anderson and K. Schwarz and L. M. Bramer and G. Z. Lomas and D. J. Orton and Z. Zhu and M. H. Engelhard and M. E. Bowden and W. C. Nelson and A. Jumpponen and J. K. Jansson and K. S. Hofmockel and C. R. Anderton},
+  title = {{A} mineral-doped micromodel platform demonstrates fungal bridging of carbon hot spots and {Hyphal} transport of mineral-derived nutrients},
   journal = {mSystems},
   volume = {7},
   number = {6},
@@ -22,7 +22,7 @@
 
 @article{Bhattacharyya2018,
   author = {A. Bhattacharyya and M. P. Schmidt and E. Stavitski and C. E. Martínez},
-  title = {Iron speciation in peats: Chemical and spectroscopic evidence for the co-occurrence of ferric and ferrous iron in organic complexes and mineral precipitates},
+  title = {{Iron} speciation in peats:~{Chemical} and spectroscopic evidence for the co-occurrence of ferric and ferrous iron in organic complexes and mineral precipitates},
   journal = {Organic Geochemistry},
   volume = {115},
   pages = {124--137},
@@ -32,7 +32,7 @@
 
 @article{Fan2021,
   author = {W. Fan and P. Yang and Z. (Joey) Yang},
-  title = {Freeze-thaw impact on macropore structure of clay by 3D X-ray computed tomography},
+  title = {{Freeze-thaw} impact on macropore structure of clay by {3D} {X-ray} computed tomography},
   journal = {Engineering Geology},
   volume = {280},
   pages = {105921},
@@ -42,7 +42,7 @@
 
 @article{Fasinmirin2018,
   author = {J. T. Fasinmirin and I. E. Olorunfemi and F. Olakuleyin},
-  title = {Strength and hydraulics characteristics variations within a tropical Alfisol in Southwestern Nigeria under different land use management},
+  title = {Strength and hydraulics characteristics variations within a tropical {Alfisol} in {Southwestern} {Nigeria} under different land use management},
   journal = {Soil and Tillage Research},
   volume = {182},
   pages = {45--56},
@@ -52,7 +52,7 @@
 
 @article{Feng2020,
   author = {Y. Feng and J. Wang and Z. Bai and L. Reading and Z. Jing},
-  title = {Three-dimensional quantification of macropore networks of different compacted soils from opencast coal mine area using X-ray computed tomography},
+  title = {Three-dimensional quantification of macropore networks of different compacted soils from opencast coal mine area using {X-ray} computed tomography},
   journal = {Soil and Tillage Research},
   volume = {198},
   pages = {104567},
@@ -73,7 +73,7 @@
 
 @article{Gostick2016,
[…]
