### 2025-08-27 6a1f662
Update to paper.bib incorporating editor's feedback
