+++ paper/paper.bib
@@ -154,7 +154,7 @@
 
 @article{Lovley1988,
   author = {D. R. Lovley and E. J. P. Phillips},
-  title = {Novel mode of microbial energy metabolism:~{Organic} carbon oxidation coupled to dissimilatory reduction of {Iron} or {Manganese},
+  title = {Novel mode of microbial energy metabolism:~{Organic} carbon oxidation coupled to dissimilatory reduction of {Iron} or {Manganese}},
   journal = {Appl Environ Microbiol},
   volume = {54},
   number = {6},
