### 2025-08-29 #8052 @kthyng
@aramyxt Thanks for signing up as a JOSS reviewer! Please take a moment to fill in your areas of expertise so we know a little more about you.
